+++ src/docfx/restore/lock/DependencyLock.cs
@@ -2,6 +2,8 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
 
 namespace Microsoft.Docs.Build
@@ -30,6 +32,30 @@ namespace Microsoft.Docs.Build
         public bool ContainsGitLock(string href)
         {
             return Git.ContainsKey(href) || Git.Keys.Any(g => g.StartsWith($"{href}#"));
+        }
+
+        public static DependencyLock Load(string docset, string dependencyLockPath)
+        {
+            Debug.Assert(!string.IsNullOrEmpty(docset));
+
+            if (string.IsNullOrEmpty(dependencyLockPath))
+            {
+                return null;
+            }
+
+            var (_, restoredLockFile) = RestoreMap.GetFileRestorePath(docset, dependencyLockPath);
+
+            // todo: add process lock
+            return JsonUtility.Deserialize<DependencyLock>(File.ReadAllText(restoredLockFile));
+        }
+
+        public static DependencyLock Load(string docset, CommandLineOptions commandLineOptions)
+        {
+            Debug.Assert(!string.IsNullOrEmpty(docset));
+
+            var (errors, config) = ConfigLoader.TryLoad(docset, commandLineOptions);
+
+            return Load(docset, config.DependencyLock);
         }
     }
 }
