--- src/docfx/restore/lock/RestoreLock.cs
@@ -1,35 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System.Diagnostics;
-using System.IO;
-
-namespace Microsoft.Docs.Build
-{
-    internal static class RestoreLock
-    {
-        public static DependencyLock Load(string docset, string dependencyLockPath)
-        {
-            Debug.Assert(!string.IsNullOrEmpty(docset));
-
-            if (string.IsNullOrEmpty(dependencyLockPath))
-            {
-                return null;
-            }
-
-            var (_, restoredLockFile) = RestoreMap.GetFileRestorePath(docset, dependencyLockPath);
-
-            // todo: add process lock
-            return JsonUtility.Deserialize<DependencyLock>(File.ReadAllText(restoredLockFile));
-        }
-
-        public static DependencyLock Load(string docset, CommandLineOptions commandLineOptions)
-        {
-            Debug.Assert(!string.IsNullOrEmpty(docset));
-
-            var (errors, config) = ConfigLoader.TryLoad(docset, commandLineOptions);
-
-            return Load(docset, config.DependencyLock);
-        }
-    }
-}
